Tasks status update fixed
Bartek encountered a bug, which prevented the number of tasks finished
to update. It was causes by my setting a task's status to the value
returned by SlaveManager ("finished") which wasn't what Scheduler
expected ("completed") thus preventing correct calculations.

Also, SlaveManager calls progress callback in enqueueTasks with a single
task and not an array of them. So I wrapped the handler to account for
this.